データ型の見直しと、faw_readの高速化。

+++ fawutil/rgy_faw_avx512.cpp
@@ -28,7 +28,7 @@
 #include "rgy_faw.h"
 
 #if defined(_M_X64) || defined(__x86_64)
-int64_t rgy_memmem_fawstart1_avx512bw(const void *data_, const int64_t data_size) {
+size_t rgy_memmem_fawstart1_avx512bw(const void *data_, const size_t data_size) {
     return rgy_memmem_avx512_imp(data_, data_size, fawstart1.data(), fawstart1.size());
 }
 #endif
