+++ fawutil/rgy_memmem_avx2.cpp
@@ -28,6 +28,6 @@
 #define RGY_MEMMEM_AVX2
 #include "rgy_memmem.h"
 
-int64_t rgy_memmem_avx2(const void *data_, const int64_t data_size, const void *target_, const int64_t target_size) {
+size_t rgy_memmem_avx2(const void *data_, const size_t data_size, const void *target_, const size_t target_size) {
     return rgy_memmem_avx2_imp(data_, data_size, target_, target_size);
 }
